Add Global Programs (`gProgs`) and surface `coreutils`
Add a Global Programs (`gProgs`) variable to the parent `base.nix`.
The `home.nix` Home Manager child will have access to this so that the parent can pass in programs in an attribute set.

Also add to `base.nix`'s `_module.args` so it can be used through parent.

The `gProgs` attribute set is split out in `home.nix`'s `_module.args`.
Each program of the attribute set then needs to be passed in / referenced indivually in other input modules.